requirements.txt
streamlit
tweepy 
joblib 
pandas 
numpy 
scikit-learn 
yfinance 
vaderSentiment 
matplotlib
python-dotenv
seaborn
imbalanced-learn
google-cloud-storage
